添加一些范例

+++ HeCommunicate/HBigProtocol.cpp
@@ -98,7 +98,7 @@ HErrorType HBigProtocol::getData(HActionType action, uint &value, int delay)
     auto error = getData(action, data, delay);
     if (error != E_OK)
         return error;
-    value = (data[0] << 24) + (data[1] << 16) + (data[2] << 8) + data[3];
+    value = static_cast<uint>((data[0] << 24) + (data[1] << 16) + (data[2] << 8) + data[3]);
     return E_OK;
 }
 
@@ -132,7 +132,7 @@ HErrorType HBigProtocol::getData(HActionType action, QVector<uint> &value, int d
         value.resize(data.size() / 4);
     auto size = qMin(value.size(), data.size() / 4);
     for (int i = 0; i < size; i++)
-        value[i] = (data[4 * i] << 24) + (data[4 * i + 1] << 16) + (data[4 * i + 2] << 8) + data[4 * i + 3];
+        value[i] = static_cast<uint>((data[4 * i] << 24) + (data[4 * i + 1] << 16) + (data[4 * i + 2] << 8) + data[4 * i + 3]);
     return E_OK;
 }
